Extract required components related items from components
